Extract etcd options from genericapiserver.
