+++ examples/apiserver/apiserver.go
@@ -28,6 +28,7 @@ import (
 	"k8s.io/kubernetes/pkg/apimachinery/registered"
 	"k8s.io/kubernetes/pkg/genericapiserver"
 	genericoptions "k8s.io/kubernetes/pkg/genericapiserver/options"
+	genericvalidation "k8s.io/kubernetes/pkg/genericapiserver/validation"
 	"k8s.io/kubernetes/pkg/storage/storagebackend"
 
 	// Install the testgroup API
@@ -51,7 +52,7 @@ func newStorageFactory() genericapiserver.StorageFactory {
 }
 
 func NewServerRunOptions() *genericoptions.ServerRunOptions {
-	serverOptions := genericoptions.NewServerRunOptions()
+	serverOptions := genericoptions.NewServerRunOptions().WithEtcdOptions()
 	serverOptions.InsecurePort = InsecurePort
 	return serverOptions
 }
@@ -61,7 +62,8 @@ func Run(serverOptions *genericoptions.ServerRunOptions) error {
 	_, serviceClusterIPRange, _ := net.ParseCIDR("10.0.0.0/24")
 	serverOptions.ServiceClusterIPRange = *serviceClusterIPRange
 	serverOptions.StorageConfig.ServerList = []string{"http://127.0.0.1:4001"}
-	genericapiserver.ValidateRunOptions(serverOptions)
+	genericvalidation.ValidateRunOptions(serverOptions)
+	genericvalidation.VerifyEtcdServersList(serverOptions)
 	config := genericapiserver.NewConfig(serverOptions)
 	config.Serializer = api.Codecs
 	s, err := genericapiserver.New(config)
