+++ pkg/genericapiserver/options/server_run_options.go
@@ -39,7 +39,6 @@ import (
 )
 
 const (
-	DefaultEtcdPathPrefix           = "/registry"
 	DefaultDeserializationCacheSize = 50000
 
 	// TODO: This can be tightened up. It still matches objects named watch or proxy.
@@ -119,10 +118,6 @@ func NewServerRunOptions() *ServerRunOptions {
 		CertDirectory:           "/var/run/kubernetes",
 		DefaultStorageMediaType: "application/json",
 		DefaultStorageVersions:  registered.AllPreferredGroupVersions(),
-		StorageConfig: storagebackend.Config{
-			Prefix: DefaultEtcdPathPrefix,
-			DeserializationCacheSize: DefaultDeserializationCacheSize,
-		},
 		DeleteCollectionWorkers: 1,
 		EnableLogsSupport:       true,
 		EnableProfiling:         true,
@@ -139,6 +134,14 @@ func NewServerRunOptions() *ServerRunOptions {
 		ServiceNodePortRange:    DefaultServiceNodePortRange,
 		StorageVersions:         registered.AllPreferredGroupVersions(),
 	}
+}
+
+func (o *ServerRunOptions) WithEtcdOptions() *ServerRunOptions {
+	o.StorageConfig = storagebackend.Config{
+		Prefix: DefaultEtcdPathPrefix,
+		DeserializationCacheSize: DefaultDeserializationCacheSize,
+	}
+	return o
 }
 
 // StorageGroupsToEncodingVersion returns a map from group name to group version,
@@ -212,7 +215,7 @@ func (s *ServerRunOptions) NewSelfClient() (clientset.Interface, error) {
 }
 
 // AddFlags adds flags for a specific APIServer to the specified FlagSet
-func (s *ServerRunOptions) AddFlags(fs *pflag.FlagSet) {
+func (s *ServerRunOptions) AddUniversalFlags(fs *pflag.FlagSet) {
 	// Note: the weird ""+ in below lines seems to be the only way to get gofmt to
 	// arrange these text blocks sensibly. Grrr.
 
@@ -300,10 +303,6 @@ func (s *ServerRunOptions) AddFlags(fs *pflag.FlagSet) {
 	// TODO: enable cache in integration tests.
 	fs.BoolVar(&s.EnableWatchCache, "watch-cache", s.EnableWatchCache,
 		"Enable watch caching in the apiserver")
-
-	fs.StringSliceVar(&s.EtcdServersOverrides, "etcd-servers-overrides", s.EtcdServersOverrides, ""+
-		"Per-resource etcd servers overrides, comma separated. The individual override "+
-		"format: group/resource#servers, where servers are http://ip:port, semicolon separated.")
 
 	fs.IntVar(&s.TargetRAMMB, "target-ram-mb", s.TargetRAMMB,
 		"Memory limit for apiserver in MB (used to configure sizes of caches, etc.)")
@@ -404,24 +403,6 @@ func (s *ServerRunOptions) AddFlags(fs *pflag.FlagSet) {
 
 	fs.StringVar(&s.StorageConfig.Type, "storage-backend", s.StorageConfig.Type,
 		"The storage backend for persistence. Options: 'etcd2' (default), 'etcd3'.")
-
-	fs.StringSliceVar(&s.StorageConfig.ServerList, "etcd-servers", s.StorageConfig.ServerList,
-		"List of etcd servers to connect with (http://ip:port), comma separated.")
-
-	fs.StringVar(&s.StorageConfig.Prefix, "etcd-prefix", s.StorageConfig.Prefix,
-		"The prefix for all resource paths in etcd.")
-
-	fs.StringVar(&s.StorageConfig.KeyFile, "etcd-keyfile", s.StorageConfig.KeyFile,
-		"SSL key file used to secure etcd communication.")
-
-	fs.StringVar(&s.StorageConfig.CertFile, "etcd-certfile", s.StorageConfig.CertFile,
-		"SSL certification file used to secure etcd communication.")
-
-	fs.StringVar(&s.StorageConfig.CAFile, "etcd-cafile", s.StorageConfig.CAFile,
-		"SSL Certificate Authority file used to secure etcd communication.")
-
-	fs.BoolVar(&s.StorageConfig.Quorum, "etcd-quorum-read", s.StorageConfig.Quorum,
-		"If true, enable quorum read.")
 
 	fs.IntVar(&s.StorageConfig.DeserializationCacheSize, "deserialization-cache-size", s.StorageConfig.DeserializationCacheSize,
 		"Number of deserialized json objects to cache in memory.")
